move special method to api
