+++ splink/database_api.py
@@ -1,4 +1,5 @@
 import logging
+from tempfile import TemporaryDirectory
 
 import duckdb
 import pandas as pd
@@ -185,3 +186,17 @@ class DuckDBAPI(DatabaseAPI):
         except ImportError:
             pass
         return accepted_df_dtypes
+
+    # special methods for use:
+
+    def export_to_duckdb_file(self, output_path, delete_intermediate_tables=False):
+        """
+        https://stackoverflow.com/questions/66027598/how-to-vacuum-reduce-file-size-on-duckdb
+        """
+        if delete_intermediate_tables:
+            self._delete_tables_created_by_splink_from_db()
+        with TemporaryDirectory() as tmpdir:
+            self._con.execute(f"EXPORT DATABASE '{tmpdir}' (FORMAT PARQUET);")
+            new_con = duckdb.connect(database=output_path)
+            new_con.execute(f"IMPORT DATABASE '{tmpdir}';")
+            new_con.close()
