+++ splink/duckdb/linker.py
@@ -2,9 +2,7 @@ from __future__ import annotations
 
 import logging
 import os
-from tempfile import TemporaryDirectory
 
-import duckdb
 import pandas as pd
 from duckdb import DuckDBPyConnection
 
@@ -157,15 +155,3 @@ class DuckDBLinker(Linker):
                 "columns. Try converting dataframes "
                 "to pyarrow tables before adding to your linker."
             ) from e
-
-    def export_to_duckdb_file(self, output_path, delete_intermediate_tables=False):
-        """
-        https://stackoverflow.com/questions/66027598/how-to-vacuum-reduce-file-size-on-duckdb
-        """
-        if delete_intermediate_tables:
-            self._delete_tables_created_by_splink_from_db()
-        with TemporaryDirectory() as tmpdir:
-            self._con.execute(f"EXPORT DATABASE '{tmpdir}' (FORMAT PARQUET);")
-            new_con = duckdb.connect(database=output_path)
-            new_con.execute(f"IMPORT DATABASE '{tmpdir}';")
-            new_con.close()
